+++ features/ui/Button/index.ts
@@ -1,1 +1,1 @@
-export { Button } from "./abutton";
+export { Button } from "./button";
